# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Buketti backend",
    "description": "Buketti backend REST API"
  },
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/editorContents": {
      "get": {
        "tags": ["EditorContents"],
        "summary": "Get all editor contents",
        "description": "Get all editor contents",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EditorContentsResponse"
            }
          }
        }
      },
      "post": {
        "consumes": ["application/json", "text/plain"],
        "tags": ["EditorContents"],
        "summary": "Create new editor content",
        "description": "Create new editor content",
        "parameters": [
          {
            "name": "editor content",
            "in": "body",
            "description": "Editor content that we want to create",
            "schema": {
              "$ref": "#/definitions/NewOrUpdatedEditorContent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New editor content is created",
            "schema": {
              "$ref": "#/definitions/EditorContentResponse"
            }
          }
        }
      }
    },
    "/api/editorContents/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Id of the editor content that we want to find",
          "type": "integer"
        }
      ],
      "get": {
        "tags": ["EditorContents"],
        "summary": "Get one editor content with given id",
        "description": "Get one editor content with given id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EditorContentResponse"
            }
          }
        }
      },
      "put": {
        "consumes": ["application/json", "text/plain"],
        "summary": "Update editor content with given id",
        "description": "Update editor content with given id",
        "tags": ["EditorContents"],
        "parameters": [
          {
            "name": "editor content",
            "in": "body",
            "description": "Updated editor content with new content",
            "schema": {
              "$ref": "#/definitions/NewOrUpdatedEditorContent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Editor content is updated",
            "schema": {
              "$ref": "#/definitions/EditorContentResponse"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete editor content with given id",
        "description": "Delete editor content with given id",
        "tags": ["EditorContents"],
        "responses": {
          "200": {
            "description": "Editor content is deleted",
            "schema": {
              "$ref": "#/definitions/EditorContentResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EditorContent": {
      "properties": {
        "id": {
          "type": "integer",
          "uniqueItems": true
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "content": {
          "type": "string"
        },
        "userId": {
          "type": "integer"
        }
      }
    },
    "EditorContentResponse": {
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "object",
          "$ref": "#/definitions/EditorContent"
        }
      }
    },
    "NewOrUpdatedEditorContent": {
      "properties": {
        "content": {
          "type": "string"
        }
      }
    },
    "EditorContentsResponse": {
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EditorContent"
          }
        }
      }
    },
...
  }
}